Struct makai::mongo::MeasurementDecimator

struct MeasurementDecimator {
    v_measurement: Option<MeasurementStatistics>,
    f_measurement: Option<MeasurementStatistics>,
    thd_measurement: Option<MeasurementStatistics>,
    pub last_insert: DateTime<Utc>,
}

Decimator for a single device.

Fields
v_measurement: Option<MeasurementStatistics>

Voltage statistics.

f_measurement: Option<MeasurementStatistics>

Frequency statistics.

thd_measurement: Option<MeasurementStatistics>

THD statistics.

last_insert: DateTime<Utc>

Last time a bson document was generated.

Methods

impl MeasurementDecimator

pub fn new() -> MeasurementDecimator

Created a new decimator.

fn clear(&mut self)

Clears the buffers.

pub fn process_message(&mut self, msg: &TriggerMessage)

Processes the next message.

Arguments:

msg- A new triggering message from a box with matching id.

pub fn generate_document_and_reset(&mut self) -> Document

Generates a bson document based on the values returned, clears the internal statistics, and resets the clock.